+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * feat: components-starter-object-storage增加腾讯cos支持
 * feat: components-starter-object-storage增加多数据源支持 @DSObjectStorage
 * feat: @SProcessor 支持处理类本身实现断言ProcessorAssert接口
+* feat: components-starter-feign-protostuff 增加`入参`支持protostuff序列化
 * feat: components-starter-feign-protostuff 增加FeignClient返回值自适应处理
 * feat: components-starter-feign-protostuff 增加请求头传递